# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service C - API Keanggotaan & Voucher",
        "description": "Service Smart Parking untuk mengelola data keanggotaan dan voucher parkir. Dibuat oleh Dinda Juniar (102022400023).",
        "contact": {
            "name": "Dinda Juniar",
            "email": "dinda@mail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/v1/memberships": {
            "get": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Melihat daftar seluruh member",
                "operationId": "eaf4ebb907ef7726f6f19c8fdda52d5e",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Mendaftarkan member baru",
                "operationId": "f3d5a078dd21f5146d6cddaf5838cc8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "membership_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rani Putri"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "rani@mail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567899"
                                    },
                                    "membership_type": {
                                        "type": "string",
                                        "example": "perak",
                                        "enum": [
                                            "perunggu",
                                            "perak",
                                            "emas",
                                            "platina"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anggota berhasil didaftarkan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/memberships/{id}": {
            "get": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Mengecek detail dan status aktif seorang member",
                "operationId": "cb9af727af4ffaafb8d1edd379452f00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member code (contoh: MEM001)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MEM001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/memberships/{id}/record-usage": {
            "post": {
                "tags": [
                    "Keanggotaan"
                ],
                "summary": "Mencatat riwayat penggunaan manfaat keanggotaan",
                "operationId": "553d07c41d2b31b47472187db92db5b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member code (contoh: MEM001)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "MEM001"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "TRX001"
                                    },
                                    "voucher_code": {
                                        "type": "string",
                                        "example": "WELCOME50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Riwayat penggunaan berhasil dicatat"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/vouchers": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Melihat daftar seluruh voucher",
                "operationId": "5e4c4903e29f39c51311fbd60882bfdd",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/validate": {
            "post": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Memvalidasi kode voucher",
                "operationId": "37c930c069799a65008528c6b95ebbc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "WELCOME50"
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "example": 10000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voucher valid"
                    },
                    "400": {
                        "description": "Voucher tidak valid"
                    },
                    "404": {
                        "description": "Voucher tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/vouchers/{code}/use": {
            "post": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Menandai voucher sebagai sudah digunakan",
                "operationId": "d5f333b120a97cd6117aa24626f3a92c",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Kode voucher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "WELCOME50"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voucher berhasil ditandai sebagai sudah digunakan"
                    },
                    "400": {
                        "description": "Voucher sudah digunakan"
                    },
                    "404": {
                        "description": "Voucher tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terotorisasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Membership": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "member_code": {
                        "type": "string",
                        "example": "MEM001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "email": {
                        "type": "string",
                        "example": "budi@mail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "membership_type": {
                        "type": "string",
                        "example": "emas",
                        "enum": [
                            "perunggu",
                            "perak",
                            "emas",
                            "platina"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif",
                        "enum": [
                            "aktif",
                            "kedaluwarsa"
                        ]
                    },
                    "discount_percent": {
                        "type": "integer",
                        "example": 20
                    },
                    "registered_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expired_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MembershipUsage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transaction_id": {
                        "type": "string",
                        "example": "TRX001"
                    },
                    "voucher_code": {
                        "type": "string",
                        "example": "WELCOME50"
                    },
                    "used_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Voucher": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "WELCOME50"
                    },
                    "description": {
                        "type": "string",
                        "example": "Diskon sambutan 50%"
                    },
                    "discount_type": {
                        "type": "string",
                        "example": "persen",
                        "enum": [
                            "persen",
                            "nominal"
                        ]
                    },
                    "discount_value": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "max_discount": {
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    },
                    "is_used": {
                        "type": "boolean",
                        "example": false
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "NIM Mahasiswa sebagai API Key (contoh: 102022400023)",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Keanggotaan",
            "description": "Keanggotaan"
        },
        {
            "name": "Voucher",
            "description": "Voucher"
        }
    ]
}